move log_utils to metrics and update imports
